Moved measurement_paths function
