+++ datacube/storage/storage.py
@@ -396,18 +396,6 @@ def _choose_location(dataset: Dataset) -> str:
     return uris[0]
 
 
-def measurement_paths(dataset: Dataset) -> Dict[str, str]:
-    """
-    Returns a dictionary mapping from band name to url pointing to band storage
-    resource.
-
-    :return: Band Name => URL
-    """
-    base = _choose_location(dataset)
-    return dict((k, _resolve_url(base, m.get('path', '')))
-                for k, m in dataset.measurements.items())
-
-
 def create_netcdf_storage_unit(filename,
                                crs, coordinates, variables, variable_params, global_attributes=None,
                                netcdfparams=None):
